% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{fitted.mylm}
\alias{fitted.mylm}
\title{Fitted values}
\usage{
\method{fitted}{mylm}(object, ...)
}
\arguments{
\item{object}{object of class "mylm"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Fitted values
}
