% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_multiply}
\alias{matrix_multiply}
\title{Matrix Multiplication}
\usage{
matrix_multiply(a, b, method = c("default", "loop", "c"))
}
\arguments{
\item{a}{First matrix}

\item{b}{Second matrix}

\item{method}{How should the multiplication be done?}
}
\description{
If method='default' then the %*% operator is used
otherwise 'loop' runs an R loop, and 'c' does the same in
compiled C code
}
