% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_multiply_c}
\alias{matrix_multiply_c}
\title{Matrix multiplication}
\usage{
matrix_multiply_c(a, b)
}
\arguments{
\item{a}{First matrix}

\item{b}{Second matrix}
}
\description{
Matrix multiplication
}
