% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylm_code.R
\name{my_lm}
\alias{my_lm}
\title{Run a basic linear regression model using my version of the lm function as a package}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula with the selected response and predicted variables to run the desired linear regression model}

\item{data}{specified datset from which the model will be ran on}
}
\value{
various information on the linear regression model such as coefficient estimates, p-values, and R-squared
}
\description{
Run a basic linear regression model using my version of the lm function as a package
}
\examples{
my_model <- my_lm(Petal.Width ~ Sepal.Length + Sepal.Width + Petal.Length, iris)
# to view information on the linear regression model otutput
my_model$Coefficients
my_model$r.squared
my_model$adj.r.squared
my_model$fstatistic
my_model$F_statistic_P_Value
head(my_model$Fitted_Values)
head(my_model$Residuals)
}
