% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{mylm}
\alias{mylm}
\title{Fit a linear model}
\usage{
mylm(formula, data, subset = NULL)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class) a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}
}
\description{
Fit a linear model
}
