% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{predict.mylm}
\alias{predict.mylm}
\title{Predicted values}
\usage{
\method{predict}{mylm}(
  object,
  newdata = NULL,
  se.fit = FALSE,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{object of class "mylm"}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{se.fit}{A switch indicating if standard errors are required (default = FALSE).}

\item{interval}{Type of interval calculation, can be "none", "confidence" or "prediction". Can be abbreviated.}

\item{level}{The confidence level required (default = 0.95).}

\item{...}{additional arguments to be passed to methods}
}
\description{
Predicted values
}
