% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{print.mylm}
\alias{print.mylm}
\title{Print method for mylm}
\usage{
\method{print}{mylm}(x, ...)
}
\arguments{
\item{x}{object of class "mylm"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Print method for mylm
}
