% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{print.summary.mylm}
\alias{print.summary.mylm}
\title{Print method for summary(mylm)}
\usage{
\method{print}{summary.mylm}(mylmobj_summary)
}
\arguments{
\item{mylmobj_summary}{Summary object}
}
\description{
Print method for summary(mylm)
}
