% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{residuals.mylm}
\alias{residuals.mylm}
\title{Residuals}
\usage{
\method{residuals}{mylm}(object, ...)
}
\arguments{
\item{object}{object of class "mylm"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Residuals
}
