% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowSumsC}
\alias{rowSumsC}
\title{Row Sums}
\usage{
rowSumsC(x)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric, complex, integer or logical values, or a numeric data frame.}
}
\description{
Row Sums
}
