% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_my_lm.R
\name{summary.my_lm}
\alias{summary.my_lm}
\title{Summary Method for my_lm}
\usage{
\method{summary}{my_lm}(object)
}
\arguments{
\item{object}{An object of class \code{"lm"} returned by \code{my_lm}.}
}
\value{
A summary of the fitted model including coefficients and residual standard error.
}
\description{
Provides a simple summary for models fitted using \code{my_lm}.
}
\examples{
data(mtcars)
fit <- my_lm(mpg ~ wt + cyl, data = mtcars)
summary(fit)
}
