% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{summary.mylm}
\alias{summary.mylm}
\title{Summary method for mylm}
\usage{
\method{summary}{mylm}(object, ...)
}
\arguments{
\item{object}{object of class "mylm"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Summary method for mylm
}
