% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{vcov.mylm}
\alias{vcov.mylm}
\title{Variance covariance matricx for parameters}
\usage{
\method{vcov}{mylm}(mylmobj)
}
\arguments{
\item{mylmobj}{mylm object}
}
\description{
Variance covariance matricx for parameters
}
