% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\docType{data}
\name{x}
\alias{x}
\title{Randomly generated x for linear regression modeling}
\format{
A matrix with 10000 rows and 5 variables generated for the normal distribution with mean = 0 and sd = 1:
}
\usage{
x
}
\description{
A dataset containing the randomly generated x.
}
\keyword{datasets}
