% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y.R
\docType{data}
\name{y}
\alias{y}
\title{Randomly generated y for linear regression modeling}
\format{
A 1-D numeric vector with 10000 rows generated by multiplying x with randomly generated betas from the normal distribution with mean = 1 and sd = 10, then adding errors randomly generated from the normal distribution with mean = 0 and sd = 0.5:
}
\usage{
y
}
\description{
A dataset containing the randomly generated y.
}
\keyword{datasets}
