% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{confint.mylm}
\alias{confint.mylm}
\title{Confidence intervals for parameters}
\usage{
\method{confint}{mylm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{object of class "mylm"}

\item{parm}{A specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{The confidence level required (default = 0.95).}

\item{...}{additional arguments to be passed to methods}
}
\description{
Confidence intervals for parameters
}
