% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fit a simple linear regression model using ordinary least squares}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Model formula, e.g., y ~ x1 + x2.}

\item{data}{Data frame containing variables.}
}
\value{
An object of class "my_lm".
}
\description{
Simplified OLS estimator based on beta_hat = (X^T X)^{-1} X^T y.
}
\examples{
dat = mtcars
fit = my_lm(mpg ~ wt + hp, data = dat)
fit$coefficients

}
