% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylm_code.R
\name{my_lm}
\alias{my_lm}
\title{Run a basic linear regression model using my version of the lm function as a package}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula with the selected response and predicted variables to run the desired linear regression model}

\item{data}{specified datset from which the model will be ran on}
}
\value{
various information on the linear regression model such as coefficient estimates, p-values, and R-squared
}
\description{
Run a basic linear regression model using my version of the lm function as a package
}
\examples{
library("sas7bdat")
data1 <- read.sas7bdat("originaldata-1.sas7bdat")
my_model <- my_lm(Depression ~ Age + Sex + R_E + Fatalism + Fatalism*Age, data1)

}
