% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLM.R
\name{myLM}
\alias{myLM}
\title{myLM}
\usage{
myLM(
  formula,
  data = NULL,
  weights = NULL,
  subset = NULL,
  na.actions = NULL,
  method = "qr",
  model = TRUE,
  x = FALSE,
  y = FALSE,
  qr = TRUE,
  contrasts = NULL
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. A 1-D vector of response is expected.}

\item{data}{an optional data frame, list or environment containing the variables in the model. Default to NULL and the variables are taken from formula.}

\item{weights}{an optional vector of weights to be used in the fitting process, default to NULL. If non-NULL, weighted least squares is used with weights weights, otherwise ordinary least squares is used.}

\item{subset}{an optional string specifying a subset of observations to be used in the fitting process, default to NULL.}

\item{na.actions}{an user specified action when dealing with NA values, can be viewed by getOption("na.action").}

\item{method}{the method to be used when fitting the model, defaulted and currently only supported with "qr". method = "model.frame" returns the model frame without fitting the model.}

\item{model, x, y, qr}{logicals. Will return the corresponding components of the fit if TRUE.}

\item{contrasts}{an optional list, whose entries are values to be used as replacement values for the contrasts replacement function and whose names are the names of columns of data containing factors.}
}
\value{
lm returns an object of class "lm". If not assigned to anything, myLM will print the function and the coefficients of the model fitted.

An object of class "lm" is a list containing at least the following components:

\item{coefficients}{a named vector of coefficients.}
\item{residuals}{a named vector of residuals.}
\item{effect}{a named vector of effects that are the uncorrelated single-degree-of-freedom values obtained by projecting the data onto the successive orthogonal subspaces generated by the QR decomposition during the fitting process.}
\item{rank}{the numeric rank of the fitted linear model.}
\item{fitted.values}{the fitted mean values.}
\item{assign}{an integer vector with an entry for each column in the matrix giving the term in the formula which gave rise to the column.}
\item{qr}{if requested, the QR decomposition of the design matrix.}
\item{qr}{if requested, the QR decomposition of the design matrix.}
\item{qr}{if requested, the QR decomposition of the design matrix.}
\item{df.residual}{the residual degrees of freedom.}
\item{call}{the matched call.}
\item{terms}{the terms object for the formula.}
\item{contrasts}{(only where relevant) the contrasts used.}
\item{model}{if requested, the model frame used, including the intersections.}
\item{x}{if requested, the design matrix used.}
\item{y}{if requested, the response used.}
\item{na.action}{(only where relevant) if not default, the assigned na.action}
}
\description{
Implementing a Linear Regression Model
}
\examples{
data("iris")
x = iris$Petal.Width
y = iris$Petal.Length
myLM(y~x)
# Equivalently
myLM(Petal.Length~Petal.Width, data = iris)

}
