% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylm.R
\name{mylm}
\alias{mylm}
\title{mylm}
\usage{
mylm(obj, inputdata = NULL, style = "simple")
}
\arguments{
\item{obj}{input formula}
}
\value{
the summary of the model
}
\description{
Fit the linear regression model
}
\examples{
mylm( mpg~wt+gear,inputdata = mtcars, style = "summary" )

}
