% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{plot.mylm}
\alias{plot.mylm}
\title{Plotting}
\usage{
\method{plot}{mylm}(mylmobj, ...)
}
\arguments{
\item{mylmobj}{mylm object}

\item{...}{additional arguments to be passed to methods, such as graphical parameters such as \code{par}.}
}
\description{
Plotting
}
