% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{predict.mylm}
\alias{predict.mylm}
\title{Predicted values}
\usage{
\method{predict}{mylm}(
  mylmobj,
  newdata = NULL,
  se.fit = FALSE,
  interval = c("none", "confidence", "prediction"),
  level = 0.95
)
}
\description{
Predicted values
}
