% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{print.summary.mylm}
\alias{print.summary.mylm}
\title{Print method for summary(mylm)}
\usage{
\method{print}{summary.mylm}(x, ...)
}
\arguments{
\item{x}{object of class "summary.mylm"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Print method for summary(mylm)
}
