% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{vcov.mylm}
\alias{vcov.mylm}
\title{Variance covariance matricx for parameters}
\usage{
\method{vcov}{mylm}(object, ...)
}
\arguments{
\item{object}{object of class "mylm"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Variance covariance matricx for parameters
}
